Fix testrace issues, move correctness tests into a new file with race testing disabled
